requirements.txt
Pillow 
requests